requirements.txt
Pillow 
requests